+++ modules/sd_detect.py
@@ -80,6 +80,8 @@ def guess_by_name(fn, current_guess):
         new_guess = 'Meissonic'
     elif 'omnigen2' in fn.lower():
         new_guess = 'OmniGen2'
+    elif 'boogu' in fn.lower():
+        new_guess = 'Boogu'
     elif 'omnigen' in fn.lower():
         new_guess = 'OmniGen'
     elif 'sd3' in fn.lower():